refactor(tab-lifecycle): split tab-lifecycle.ts into per-concern submodules
Replace agent/tab-lifecycle.ts (708 LOC) with a tab-lifecycle/ directory:

- utils.ts          - modelKey / modelDescriptor / compilePattern /
                      tabSessionDir + TabLifecycleDeps interface
- terminal.ts       - emitBashResult (terminal panel chunked emit)
- tools.ts          - cancelRunningToolCards (reload-sentinel synthetic
                      end-state; preserves the "extension drops never
                      abort user's LLM turn" guarantee)
- models.ts         - buildPickerModels / defaultModelKey /
                      ensurePickerHasModel
- slash-commands.ts - collectPiSlashCommands + refreshPiSlashCommands
                      (extension / prompt / skill discovery)
- ready-handshake.ts - emitReady (startup payload to frontend)
- events.ts         - handleSessionEvent (the per-turn state machine:
                      agent_start/end, message_update, tool_execution_*)
- lifecycle.ts      - ensureTab + EnsureTabOptions (per-tab session
                      creation, subscriber wiring)
- index.ts          - barrel re-exporting the full public surface

Public surface unchanged. The 7 external consumers (dispatcherTypes,
main, a2uiEvents, chat, nativeSlash, tabs, tab-lifecycle.test.ts) keep
importing from "./tab-lifecycle" which now resolves to the directory's
index.ts. The reload-sentinel test coverage in tab-lifecycle.test.ts
remains green (1124/1124 vitest pass, including the
"cancelRunningToolCards then tool_execution_end" deduplication test).

Also bump eslint.config.js's projectService allowDefaultProject to
match agent/*/*.ts (one level of nesting; tseslint disallows **) and
raise the default-project file-match cap from 64 to 128 to absorb the
new submodules.
